+++ src/app/mate/[id]/components/MateInfo/components/index.ts
@@ -1,2 +1,2 @@
-export { default as MateToggleMenu } from "./MateToggleMenu/MateToggleMenu";
+export { default as MateDropDownMenu } from "./MateDropDownMenu/MateDropDownMenu";
 export { default as MateActions } from "./MateActions/MateActions";
